Move to a separate module and fully test
